backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==3.0.1
email-validator==2.1.0
Pillow==10.1.0
gunicorn==21.2.0
python-slugify==8.0.1